requirements.txt
flask
openai
langchain_openai
langchain_community
faiss-cpu
pypdf
neo4j
langchain_experimental
networkx
matplotlib
langgraph
langchain_neo4j
pandas
janome
unstructured
python-docx
python-pptx
openpyxl